Author: R. Mitchell Changes: 1. Confidence: Added a confidence threshold parameter to display the minimum cue confidence required for successful menotaxis. This gives some graphical indication of confidence without making the beetle's direction unclear.
2. Jupyter Notebook:
Added a basic but functional Jupyter notebook which allows a user to
easily edit the config file and run the software to get an interactive plot.
There are still issues with the functionality, the config is set up as a code
cell so that it runs properly, however, that also means I get python errors when
running it. It works, but I need a fix for this. This required a slight
modification to the scratch file but this file should be replaced before
'release' in any case.

Testing:
1. Checked the plots were correctly displaying the confidence and there were
no errors if the parameter wasn't specified. (i.e. that the default is correctly
used).

2. Checked that the config update works and produces the correct plots when
modified.