Author: Robert Mitchell Changes: Updated plots so that each column shows only the plots for their respective rolls. This should make the plots clearer to read. Updated legends accordingly so that they reflect the change.
There is still code in the Beetle class that could be changed to make
operation more streamlined but this was deemed inconsequential. It
doesn't change how the software runs or how easy it is to use.

Testing:
Ran the software with default configuration and inspected the plots to
ensure they were correct.