Author: R. Mitchell Changes: Modified the effect that the linear polarising cue has on the avg combined cue.
- Rather than presenting two equal and opposite vectors the sensory
  vector for each direction is treated separately.
- We compute the combined cue using the master and the mirror
  separately.
- The one which gives greatest cue confidence is used.
- If neither is greater, choose at random. (e.g. light at azimuth
  zero, polarisation along the -90,90 axis)

How sensible this is requires further discussion but it seems like the
most obvious way of integrating polarisation into the avg
strategy. Another approach would be to take the beetle's starting
bearing into account, however this requires polarisation to be a
dominant cue and is therefore better suited to a wta strategy.

If we take El Jundi et al. 2015 (CX Coding paper) it seems that the
beetles have a solid preference and ignore the non-dominant cue. In
this case, polarisation would first have to be the dominant cue, then the
beetle would follow whichever of the sensory vectors which are closest
to its initial bearing. Thus, beetle bearing will need to be taken
into account when building wta.

Testing:
Wrote an untracked config file with polarisation and a
light-point. Varied the relative position of each cue to check that
the chosen polarisation direction fit within the strategy.