Author: R. Mitchell Changes: Documentation updates. Updated the README to be slightly more clear in the installation instructions. Updated the General notebook to include some information on Polarisation and an attempt to clarify the difference between cue strength and confidence. The console output image has also been updated to reflect new options.
Code changes. I noticed that the terminal output was split by the plot. This
wasn't noticed as plots didn't used to be included in the notebook. I've updated
this so that all of the output is together and matches the new picture.

Testing:
Only functional change was the notebook terminal output, made sure this was being
displayed correctly.